requirements.txt
Flask
gunicorn
Pillow
kraken